ci pipeline and test  update

+++ src/features/chats/application/use_cases/delete.py
@@ -1,6 +1,7 @@
 from uuid import UUID
-from src.persistence import DataRepository
-from ...domain import ChatPublic
+from src.persistence import DataRepository, NotFoundException
+from src.security import PermissionsException
+from ...domain import ChatPublic, Chat
 
 class DeleteChat:
     def __init__(
@@ -11,8 +12,19 @@ class DeleteChat:
 
     def execute(
         self,
+        user_id: UUID,
         chat_id: UUID
     ):
+        chat: Chat = self.__chat_repository.get_one(
+            key="chat_id",
+            value=chat_id
+        )
+
+        if not chat:
+            raise NotFoundException()
+        
+        if str(user_id) != str(chat.agent.user_id):
+            raise PermissionsException()
         
         deleted_chat = self.__chat_repository.delete(
             key="chat_id",
